refactor: move homepage from views folder to app folder
