+++ apps/site/src/app/(home)/sections/FAQ/FAQ.module.scss
