+++ apps/site/src/app/(home)/sections/FAQ/FAQ.tsx
