+++ apps/site/src/app/(home)/sections/FAQ/FAQAccordion.tsx
