+++ apps/site/src/app/(home)/sections/Sponsors/Sponsors.module.scss
