+++ apps/site/src/app/(home)/sections/Sponsors/Sponsors.tsx
