+++ apps/site/src/app/(home)/sections/Sponsors/getSponsors.ts
