--- apps/site/src/app/page.tsx
@@ -1,3 +1,0 @@
-export const revalidate = 60;
-
-export { Home as default } from "@/views";
