+++ apps/site/src/views/index.ts
@@ -1,4 +1,3 @@
-export { default as Home } from "./Home/Home";
 export { default as Maintenance } from "./Maintenance/Maintenance";
 export { default as Resources } from "./Resources/Resources";
 export { default as Schedule } from "./Schedule/Schedule";
